order_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-fastapi-instrumentator
requests==2.32.3
python-json-logger
python-dotenv
httpx
tenacity
python-logstash

